+++ api/features/workflows/core/models.py
@@ -102,6 +102,10 @@ class ChangeRequest(
 
     ignore_conflicts = models.BooleanField(default=False)
 
+    class Meta:
+        # Explicit ordering to prevent pagination warnings.
+        ordering = ("id",)
+
     def approve(self, user: "FFAdminUser"):
         if user.id == self.user_id:
             raise CannotApproveOwnChangeRequest(
@@ -277,7 +281,7 @@ class ChangeRequest(
             url += f"/project/{self.environment.project_id}"
             url += f"/environment/{self.environment.api_key}"
         else:
-            url += f"/projects/{self.project_id}"
+            url += f"/project/{self.project_id}"
         url += f"/change-requests/{self.id}"
         return url