+++ frontend/common/types/requests.ts
@@ -16,6 +16,8 @@ import {
   Identity,
   ChangeRequest,
   ProjectChangeRequest,
+  Role,
+  RolePermission,
 } from './responses'
 
 export type PagedRequest<T> = T & {
@@ -170,7 +172,7 @@ export type Req = {
   updateRole: {
     organisation_id: number
     role_id: number
-    body: { description: string | null; name: string }
+    body: Role
   }
   deleteRole: { organisation_id: number; role_id: number }
   getRolePermissionEnvironment: {
@@ -191,7 +193,7 @@ export type Req = {
     level: PermissionLevel
     body: {
       admin?: boolean
-      permissions: string[]
+      permissions: RolePermission['permissions']
       project: number
       environment: number
     }