+++ frontend/web/components/App.js
@@ -440,6 +440,7 @@ const App = class extends Component {
                                       >
                                         <NavLink
                                           id='org-settings-link'
+                                          data-test='org-settings-link'
                                           activeClassName='active'
                                           className={classNames(
                                             'breadcrumb-link',
@@ -573,6 +574,7 @@ const App = class extends Component {
                                   icon={<AuditLogIcon />}
                                   id='audit-log-link'
                                   to={`/project/${projectId}/audit-log`}
+                                  data-test='audit-log-link'
                                 >
                                   Audit Log
                                 </NavSubLink>
@@ -663,6 +665,7 @@ const App = class extends Component {
                                 <NavSubLink
                                   icon={<SettingsIcon />}
                                   id='org-settings-link'
+                                data-test='org-settings-link'
                                   to={`/organisation/${
                                     AccountStore.getOrganisation().id
                                   }/settings`}