+++ frontend/web/components/pages/ProjectSettingsPage.js
@@ -26,6 +26,7 @@ import PlanBasedBanner from 'components/PlanBasedAccess'
 import classNames from 'classnames'
 import ProjectProvider from 'common/providers/ProjectProvider'
 import ChangeRequestsSetting from 'components/ChangeRequestsSetting'
+import { getRoleProjectPermissions } from 'common/services/useRolePermission'
 
 const ProjectSettingsPage = class extends Component {
   static displayName = 'ProjectSettingsPage'
@@ -59,6 +60,20 @@ const ProjectSettingsPage = class extends Component {
       getStore(),
       { organisation_id: AccountStore.getOrganisation().id },
       { forceRefetch: true },
+    ).then((roles) => {
+      if (!roles?.data?.results?.length) return
+
+      getRoleProjectPermissions(
+        getStore(),
+        {
+          organisation_id: AccountStore.getOrganisation().id,
+          project_id: this.props.match.params.projectId,
+          role_id: roles.data.results[0].id,
+        },
+        { forceRefetch: true },
+      ).then((res) => {
+        const matchingItems = roles.data.results.filter((item1) =>
+          res.data.results.some((item2) => item2.role === item1.id),
     )
   }